Modify ContainsAny / All tests to run also in cluster setup
